requirements.txt
# Core
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
pydantic==2.6.4

# Data & APIs
pandas==2.2.1
google-genai==1.29.0


# Utilities
python-dateutil==2.9.0.post0  
requests==2.31.0
numpy==1.26.4  # Pandas dependency

# Dev/Testing (optional)
gunicorn==21.2.0  # Production WSGI server
pytest==8.0.2